+++ src/nativelib/common/dtypes/functions/enums.pyx
@@ -14,7 +14,7 @@ cdef dict EnumStructValue = {
 cpdef str read_enum(
     object fileobj,
     int length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     dict enumcase,
@@ -30,7 +30,7 @@ cpdef str read_enum(
 cpdef bytes write_enum(
     object dtype_value,
     int length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     dict enumcase,