+++ src/nativelib/common/dtypes/functions/floats.pyx
@@ -18,7 +18,7 @@ cdef dict FloatStructValue = {
 cpdef float read_bfloat16(
     object fileobj,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -53,7 +53,7 @@ cpdef float read_bfloat16(
 cpdef bytes write_bfloat16(
     object dtype_value,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -87,7 +87,7 @@ cpdef bytes write_bfloat16(
 cpdef double read_float(
     object fileobj,
     int length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -102,7 +102,7 @@ cpdef double read_float(
 cpdef bytes write_float(
     object dtype_value,
     int length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,