+++ src/nativelib/common/dtypes/functions/ipaddrs.pyx
@@ -4,7 +4,7 @@ from ipaddress import ip_address
 cpdef object read_ipv4(
     object fileobj,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -18,7 +18,7 @@ cpdef object read_ipv4(
 cpdef bytes write_ipv4(
     object dtype_value,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -35,7 +35,7 @@ cpdef bytes write_ipv4(
 cpdef object read_ipv6(
     object fileobj,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -49,7 +49,7 @@ cpdef object read_ipv6(
 cpdef bytes write_ipv6(
     object dtype_value,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,