+++ src/nativelib/common/dtypes/functions/uuids.pyx
@@ -8,7 +8,7 @@ from uuid import UUID
 cpdef object read_uuid(
     object fileobj,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,
@@ -25,7 +25,7 @@ cpdef object read_uuid(
 cpdef bytes write_uuid(
     object dtype_value,
     object length,
-    object precission,
+    object precision,
     object scale,
     object tzinfo,
     object enumcase,